+++ src/__fixtures__/operations/getCollection.ts
@@ -9,8 +9,9 @@ export const getCollectionsMock = [
       data: {
         collections: [
           {
-            _id: '34',
+            _id: '1',
             title: 'Example Collection',
+            type: 'Collection',
             bookmarks: [
               {
                 _id: '3',
@@ -38,7 +39,60 @@ export const getCollectionsMock = [
               },
             ],
           },
+          {
+            _id: '2',
+            title: 'Maxed Out Collection',
+            type: 'Collection',
+            bookmarks: Array.from({ length: 10 }, (x, i) => ({
+              _id: `${i}`,
+              label: `Bookmark ${i}`,
+              url: '#',
+              isRemoved: null,
+              cmsId: `${i}`,
+            })),
+          },
+          ...Array.from({ length: 20 }, (x, i) => ({
+            _id: `${i + 3}`,
+            title: `Collection ${i + 3}`,
+            type: 'Collection',
+            bookmarks: [],
+          })),
         ],
+      },
+    },
+  },
+]
+
+const mockCollection = {
+  title: 'Example Collection',
+  type: 'Collection',
+  bookmarks: [
+    {
+      _id: '3',
+      url: 'https://google.com',
+      label: 'Webmail',
+      description: 'Lorem ipsum',
+      cmsId: null,
+      isRemoved: null,
+    },
+  ],
+}
+
+const maxCollections = Array.from({ length: 25 }, (x, i) => {
+  return {
+    ...mockCollection,
+    _id: `${i}`,
+  }
+})
+
+export const getMaximumCollectionsMock = [
+  {
+    request: {
+      query: GET_COLLECTIONS,
+    },
+    result: {
+      data: {
+        collections: maxCollections,
       },
     },
   },