+++ src/constants/index.ts
@@ -1,0 +1,5 @@
+export const SPACEFORCE_NEWS_RSS_URL = `https://www.spaceforce.mil/DesktopModules/ArticleCS/RSS.ashx?ContentType=1&Site=1060`
+
+export const WIDGET_TYPES = {
+  COLLECTION: 'Collection',
+} as const