+++ src/schema.tsx
@@ -8,9 +8,15 @@ export const typeDefs = gql`
     cmsId: ID
     isRemoved: Boolean
   }
+
+  enum WidgetType {
+    Collection
+  }
+
   type Collection {
     _id: ID!
     title: String!
+    type: WidgetType!
     bookmarks: [Bookmark]
   }
   type Query {
@@ -28,6 +34,12 @@ export const typeDefs = gql`
     ): Bookmark
     addCollections(collections: [CollectionRecord!]): [Collection]
     removeBookmark(_id: ID!, collectionId: ID!, cmsId: ID): Bookmark
+    editBookmark(
+      _id: ID!
+      collectionId: ID!
+      url: String
+      label: String
+    ): Bookmark
   }
   input BookmarkInput {
     url: String!